requirement.txt
torch
transformers
pandas
Pillow
torchvision
taming-transformers
transformers
clip-by-openai
ftfy
regex
pillow